refactor: extract CLI recipe validation helpers
